requirements.txt
beautifulsoup4
django
djangorestframework
requests
pymysql